tweaks for 64-bit builds
git-svn-id: https://svn.r-project.org/R/trunk@42577 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/graphapp/buttons.c
@@ -875,7 +875,7 @@ void inserttext(control obj, const char *text)
 	if ((obj->kind != FieldObject) && (obj->kind != TextboxObject))
 		return;
 	text = to_dos_string(text);
-	sendmessage(obj->handle, EM_REPLACESEL, 0, (long) text);
+	sendmessage(obj->handle, EM_REPLACESEL, 0, (intptr_t) text);
 	if (text)
 		discard(text);
 }
